Crate rust_wasm_dodrio_router

rust_wasm_dodrio_router

version: 0.4.6 date: 2020-05-09 authors: Luciano Bestia
wasm router for local hash routes for dodrio vdom

local router with hash for dodrio vdom

I needed a router for local hash routes in rust wasm for dodrio vdom.
This library contains the generic parts.

In the project add a file/mod with specific implementation code like this example:
https://github.com/LucianoBestia/mem6_game/blob/master/mem6/src/routerimplmod.rs

cargo crev reviews and advisory

It is recommended to always use cargo-crev
to verify the trustworthiness of each of your dependencies.
Please, spread this info.
On the web use this url to read crate reviews. Example:
https://web.crev.dev/cargo_crev_web/crate/num-traits

Modules

router_mod
router_mod - A simple #-fragment local router for dodrio vdom and html templating This is the trait module. It is a lib crate. It does not know anything about the data model of the project. That is abstracted away with field get/set methods to implement. All the implementation for a project are isolated in the project module router_impl_mod. I couldn't abstract it away from dodrio vdom. It is still a dependency.