Module rust_wasm_dodrio_router::router_mod

router_mod - A simple #-fragment local router for dodrio vdom and html templating This is the trait module. It is a lib crate. It does not know anything about the data model of the project. That is abstracted away with field get/set methods to implement. All the implementation for a project are isolated in the project module router_impl_mod. I couldn't abstract it away from dodrio vdom. It is still a dependency.

Traits

RouterTrait
methods for Router, some required to be implemented, other with default implementation in this file Traits cannot have fields. They must use access methods for every single field. dodrio is used for event handling, so I must

Functions

private_get_url_param_in_hash_after_dot
I cannot test associated methods. So I make a normal private fn.