Trait rust_wasm_dodrio_router::router_mod::RouterTrait

pub trait RouterTrait {
    fn get_file_name_to_fetch_from_dodrio(root: &mut dyn RootRender) -> &str;

    fn set_file_name_to_fetch_from_dodrio(
        file_name_to_fetch: String, 
        root: &mut dyn RootRender, 
        vdom: VdomWeak
    ) -> String;

    fn set_fetched_file(
        resp_body_text: String
    ) -> Box<dyn Fn(&mut dyn RootRender) + 'static>;

    fn start_router(&self, vdom: VdomWeak) { ... }

    fn set_on_hash_change_callback(&self, on_hash_change: Box<dyn FnMut()>) { ... }

    fn get_url_param_in_hash_after_dot(short_route: &str) -> &str { ... }
}

methods for Router, some required to be implemented, other with default implementation in this file Traits cannot have fields. They must use access methods for every single field. dodrio is used for event handling, so I must

Required methods

fn get_file_name_to_fetch_from_dodrio(root: &mut dyn RootRender) -> &str

fn set_file_name_to_fetch_from_dodrio(
file_name_to_fetch: String,
root: &mut dyn RootRender,
vdom: VdomWeak
) -> String

fn set_fetched_file(
resp_body_text: String
) -> Box<dyn Fn(&mut dyn RootRender) + 'static>

Loading content...
Provided methods

fn start_router(&self, vdom: VdomWeak)

Start the router.

fn set_on_hash_change_callback(&self, on_hash_change: Box<dyn FnMut()>)

fn get_url_param_in_hash_after_dot(short_route: &str) -> &str

get the first param after hash in local route after dot example #p03.1234 -> 1234

Loading content...