Trait rust_wasm_router::routermod::RouterTrait

pub trait RouterTrait {
    fn get_local_route_from_self(&self) -> &str;

    fn get_local_route_from_root(root: &mut dyn RootRender) -> &str;

    fn update_local_route_from_root(
        local_route: String, 
        root: &mut dyn RootRender, 
        vdom: VdomWeak
    ) -> String;

    fn update_html_template_and_sub_templates(
        resp_body_text: String
    ) -> Box<dyn Fn(&mut dyn RootRender) + 'static>;

    fn start_router(&self, vdom: VdomWeak) { ... }

    fn set_on_hash_change_callback(&self, on_hash_change: Box<dyn FnMut()>) { ... }

    fn get_url_param_in_hash_after_dot(short_local_route: &str) -> &str { ... }
}

methods for Router, some generic, some specific Traits cannot have fields. They must use access methods for every single field.

Required methods

fn get_local_route_from_self(&self) -> &str

fn get_local_route_from_root(root: &mut dyn RootRender) -> &str

fn update_local_route_from_root(
local_route: String,
root: &mut dyn RootRender,
vdom: VdomWeak
) -> String

fn update_html_template_and_sub_templates(
resp_body_text: String
) -> Box<dyn Fn(&mut dyn RootRender) + 'static>

Loading content...
Provided methods

fn start_router(&self, vdom: VdomWeak)

Start the router.

fn set_on_hash_change_callback(&self, on_hash_change: Box<dyn FnMut()>)

fn get_url_param_in_hash_after_dot(short_local_route: &str) -> &str

get the first param after hash in local route after dot example &p03.1234 -> 1234

Loading content...