Trait rust_wasm_dodrio_router::router_mod::RouterTrait

pub trait RouterTrait {
    fn get_location_hash(&self) -> &str;

    fn get_file_name_to_fetch(&self) -> &str;

    fn set_file_name_to_fetch(
        &mut self, 
        file_name_to_fetch: String, 
        vdom: VdomWeak
    ) -> String;

    fn from_root_to_router_data(root: &mut dyn RootRender) -> &mut Self;

    fn fn_to_prepare_fetched_html_for_render(
        resp_body_text: String
    ) -> Box<dyn Fn(&mut dyn RootRender) + 'static>;

    fn start_router(&self, vdom: VdomWeak) { ... }

    fn set_on_hash_change_callback(&self, on_hash_change: Box<dyn FnMut()>) { ... }
}

methods for Router, some required to be implemented, other with default implementation in this file Traits cannot have fields. They must use access methods for every single field. dodrio is used for event handling, so I must

Required methods

fn get_location_hash(&self) -> &str

get location hash

fn get_file_name_to_fetch(&self) -> &str

get file name

fn set_file_name_to_fetch(
&mut self,
file_name_to_fetch: String,
vdom: VdomWeak
) -> String

set file name

fn from_root_to_router_data(root: &mut dyn RootRender) -> &mut Self

from root to router data

fn fn_to_prepare_fetched_html_for_render(
resp_body_text: String
) -> Box<dyn Fn(&mut dyn RootRender) + 'static>

fn to prepare fetched html for render

Loading content...
Provided methods

fn start_router(&self, vdom: VdomWeak)

Start the router.

fn set_on_hash_change_callback(&self, on_hash_change: Box<dyn FnMut()>)

Loading content...